refactor: extract key value components
