+++ src/components/molecules/FilterPopover/keyValueInput/index.ts
@@ -1,0 +1,2 @@
+export * from './KeyValueInput';
+export * from './NewKeyValueInput';
